+++ packages/@straw-hat/openapi-web-sdk-generator/bin/dev.js
@@ -2,8 +2,8 @@
 
 // eslint-disable-next-line node/shebang
 async function main() {
-    const {execute} = await import('@oclif/core')
-    await execute({development: true, dir: import.meta.url})
+  const { execute } = await import('@oclif/core');
+  await execute({ development: true, dir: import.meta.url });
 }
 
-await main()
+await main();