+++ packages/@straw-hat/prettier-config/lib/base.js
@@ -1,4 +1,4 @@
-export default {
+module.exports = {
   $schema: 'https://json.schemastore.org/prettierrc',
   printWidth: 120,
   tabWidth: 2,