+++ packages/@straw-hat/prettier-config/lib/index.js
@@ -1,7 +1,2 @@
-import baseConfig from './base.js';
-
-export default {
-  ...baseConfig,
-  organizeImportsSkipDestructiveCodeActions: true,
-  plugins: [...baseConfig.plugins, 'prettier-plugin-organize-imports'],
-};
+const baseConfig = require('./base.js');
+module.exports = baseConfig;