+++ packages/@straw-hat/types/src/css-modules.ts
@@ -1,8 +1,10 @@
+// @ts-expect-error
 declare module '*.module.css' {
   const classes: { readonly [key: string]: string };
   export default classes;
 }
 
+// @ts-expect-error
 declare module '*.module.scss' {
   const classes: { readonly [key: string]: string };
   export default classes;