+++ packages/@straw-hat/react-query-fetcher/CHANGELOG.md
@@ -7,4 +7,4 @@
 ### Breaking Changes
 
 - Requires Node.js 16.0.0 or higher. This is due to the usage of ESM modules only.
-- Requires `@tanstack/react-query@v5+` and `@straw-hat/fetcher@v5+`. 
+- Requires `@tanstack/react-query@v5+` and `@straw-hat/fetcher@v5+`.